Move `LoginService` to production
