+++ Tests/Curtz/Login/LoginServiceTests.swift
@@ -8,84 +8,6 @@
 import XCTest
 import Curtz
 
-public struct LoginMapper {
-    
-    private struct Item: Decodable {
-        let id: String
-        let email: String
-        let created_at: Date
-        let updated_at: Date
-        let access_token: String
-        let refresh_token: String
-        
-        var response: LoginResponse {
-            return LoginResponse(id: id, email: email, createdAt: created_at, updatedAt: updated_at, accessToken: access_token, refreshToken: refresh_token)
-        }
-    }
-    
-    static func map(_ data: Data, from response: HTTPURLResponse) -> LoginService.Result {
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-        
-        guard isOK(response), let res = try? decoder.decode(Item.self, from: data) else {
-            return .failure(LoginService.Error.invalidData)
-        }
-        return .success(res.response)
-        
-    }
-    
-    private static func isOK(_ response: HTTPURLResponse) -> Bool {
-        (200...299).contains(response.statusCode)
-    }
-    
-}
-
-public class LoginService {
-    private let client: HTTPClient
-    private let loginURL: URL
-    
-    public typealias Result = LoginResult
-    public enum Error: Swift.Error {
-        case connectivity
-        case invalidData
-    }
-    
-    public init(loginURL: URL, client: HTTPClient) {
-        self.client = client
-        self.loginURL = loginURL
-    }
-    
-    public func login(user: LoginRequest, completion: @escaping (Result) -> Void) {
-        let request = prepareRequest(for: user)
-        
-        client.perform(request: request) {[weak self] result in
-            guard self != nil else { return }
-            switch result {
-            case let .success((data, response)):
-                completion(LoginMapper.map(data, from: response))
-            case .failure:
-                completion(.failure(Error.connectivity))
-            }
-        }
-        
-        
-        func prepareRequest(for user: LoginRequest) -> URLRequest {
-            var request = URLRequest(url: self.loginURL)
-            request.httpMethod = "POST"
-            request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-            
-            let requestBody: [String: String] = [
-                "email": user.email,
-                "password": user.password
-            ]
-            let jsonData = try? JSONSerialization.data(withJSONObject: requestBody)
-            request.httpBody = jsonData
-            
-            return request
-        }
-    }
-}
-
 final class LoginServiceTests: XCTestCase {
     
     func test_init_doesNOTperformAnyRequest() {
